rename AllSerachResults to AllSearchComponents
