+++ src/frontend/src/containers/AllSearchComponents.tsx
